03 - Gestão de Detalhes do Plano -  aula 08 - Não Permitir Deletar Plano com Detalhes no LaraFood

+++ app/Http/Controllers/Admin/PlanController.php
@@ -55,12 +55,22 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('url', $url)->first();
 
         if(!$plan){
             return redirect()->back();
         }
-        
+
+
+
+
+        if( $plan->details->count() > 0 ){
+            return redirect()->back()->with('error','Existem detalhes vinculados a esse plano,
+            portanto nao é possível deletar!');
+
+        }
         $plan->delete();
 
         return redirect()->route('plans.index');
@@ -85,7 +95,7 @@ class PlanController extends Controller
         if(!$plan){
             return redirect()->back();
         }
-        
+
         return view('admin.pages.plans.edit', compact('plan'));
 
     }
@@ -104,6 +114,6 @@ class PlanController extends Controller
         return redirect()->route('plans.index');
 
     }
-    
+
 }
